HTTP connection handling improvements.
Previous issue of http.client.BadStatusLine was due to imgur closing the
tcp connection we used to fetch metadata. Likely this was done after the
keepalive timeout on the server.

This would be the same issue that causes Apache httpcomponents to raise
NoHttpResponseException.

Changed the script to not reuse http connections when getting metadata.

There is a slightly higher chance if this issue coming up again - when
fetching images. It's possible to add retry logic there if necessary.